requirements.txt
# Core requirements
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# Database
# psycopg2>=2.9.6  # Uncomment if using PostgreSQL

# Development
django-debug-toolbar>=4.1.0

# Testing
pytest>=7.3.1
pytest-django>=4.5.2
